Umbral Magic Suggestions and Slight Repositioning
At this point it's time we address the fact that these Soldier sprites
are way too big. There is no way we could cram 10 of these into one
fight. This cannot be the default enemy size, even for this fight.

Also gave some suggestions for how Cleaver's spells could work. You
don't have to listen to it because it doesn't really matter for the core
of the game :^) but it's there if you want