Battle Changes
Fixed the damage range on the damage formula, turns out rand actually
goes between 0 and [specified number]-1, so i added +1 to every
[specified number]

Changed the ATK on the Aged Axegun from 20 to 15, changed the damage
formula for Knives to be the same as Knife, and changed Shasho's power
from 18 to 26. There seems to be a bug with getting Shasho to have an HP
sacrifice, I'll do something about that soon.