More Battle Stuff
Changed everyone's HP to what it says in the stat doc
Changed Daraway's stats to be more like that of an actual human person
Set Deadly Blow's Hit Rate to 100, it's still missing. Every time. Is
anyone else having this problem? If you used a rand function for hit
rate, it needs to be rand(2), not rand(1)
Made it so Shasho actually has an HP cost. I listed in the tasks doc
what my problems with this are