charisma almost added
-the base for the charisma stat has been layed down. unfortunately it
does not work, as something in its definitions doesn't work out and when
it's called it crashes the game.
  def base_cha
    n =
base_parameter(6)
    n *= Actor_Mult_Cha
    for item in
equips.compact do n += item.cha_plus end
    return n.to_i
  end
-a nonworking aggro script has been added to gay baby jail for later
fixing maybe
-lightning and thunder can now be called via a script command (talk to
Fang while on the ship to see it in action)
-days of the week can be checked now! look for the enterence to the
Saturday room when variable 20 is equal to 6! (this variable is updated
after the welcome message and when talking to calhoun. it should be
updated once upon entering a map that uses it)
-longer script call script added for nicer scriptlet functionality
-grandma cat lady added
-advanced map scroling added